Update day6.java
Combined Part 1 and Part 2
Revamped to use a method to check for repeats, so that we can switch between 4 for Part 1 and 14 for Part 2.
The index for Part 1 sets the index for Part 2 so it doesn't test the beginning twice. 

Could be improved, the old version would increment based on where the repeat was located, but this version only increments by 1 every time.