Update day11.java
Updated for Part 2.

Now prompts for which part the user wants to do.

Had some help from other users on reddit, I was on the right track looking at ways to reduce the numbers but was focusing on the individual monkey actions + checks and not the overall checks.

Found out I had mis-initialized a monkey, typed 60 and 611 instead of 50 and 51. Didn't seem to change the first part, but problem compounded over 10,000 rounds.